implement read_header
